Change timeout to be of type unsigned long
We had a bug where timout becomes negative in the sendBuffer function because it's passed as a signed int. The default 60000 ms timeout overflows the range of a signed int on Arduino Uno (AVR). Int is 16 bits, thus 2^16 = 65536. Maximum values for a signed int is therefore 65536/2-1 = 32767. And we are trying to store 60000 into this. Therefore overflows to  -5535.